Renamed to CartCrosssellsScroller
