+++ packages/magento-cart-items/index.ts
@@ -1,10 +1,11 @@
 export * from './Api/CartItem.gql'
 export * from './Api/CartItems.gql'
+export * from './components/CartCrosssellsScroller/CartCrosssellsScroller'
 export * from './components/CartItem/CartItem'
-export * from './components/CartItemsActionCards/CartItemsActionCards'
 export * from './components/CartItemActionCard/CartItemActionCard'
-export * from './components/SelectedCustomizableOptions/SelectedCustomizableOptions'
+export * from './components/CartItemsActionCards/CartItemsActionCards'
 export * from './components/RemoveItemFromCart/RemoveItemFromCart'
-export * from './components/RemoveItemFromCart/RemoveItemFromCartFab'
 export * from './components/RemoveItemFromCart/RemoveItemFromCart.gql'
+export * from './components/RemoveItemFromCart/RemoveItemFromCartFab'
+export * from './components/SelectedCustomizableOptions/SelectedCustomizableOptions'
 export * from './components/UpdateItemQuantity/UpdateItemQuantity'
