+++ packages/next-ui/Overlay/components/index.ts
@@ -1,2 +1,3 @@
 export * from './OverlayBase'
 export * from './Overlay'
+export * from './OverlayStickyBottom'
